Move packagedcode.__init__.py code to handlers.py
Move packagedcode.__init__.py code to handlers.py to allow distributing
multiple smaller packages built with flot.
